testing gatt classes refactor..
